Update training-data-format.rst
For the example "Could I pay in [yen](currency)?" under "## intent:check_balance", the lookup table name for the entity to refer to was mentioned incorrectly, as it is supposed to refer to the lookup table "currencies". Hence I feel it is required to change "Could I pay in [yen](currency)?" to "Could I pay in [yen](currencies)?" to avoid confusion. This did cause me some confusion and ambiguity when I started learning how to use RASA.